refactor: extract SubsonicUserController — fixes #108
