+++ airsonic-main/src/main/java/org/airsonic/player/controller/SubsonicUserController.java
@@ -1,0 +1,295 @@
+/*
+ This file is part of Airsonic.
+
+ Airsonic is free software: you can redistribute it and/or modify
+ it under the terms of the GNU General Public License as published by
+ the Free Software Foundation, either version 3 of the License, or
+ (at your option) any later version.
+
+ Airsonic is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License
+ along with Airsonic.  If not, see <http://www.gnu.org/licenses/>.
+
+ Copyright 2025 (C) Airsonic Authors
+ Based upon Subsonic, Copyright 2009 (C) Sindre Mehus
+ */
+package org.airsonic.player.controller;
+
+import org.airsonic.player.command.UserSettingsCommand;
+import org.airsonic.player.domain.TranscodeScheme;
+import org.airsonic.player.domain.UserSettings;
+import org.airsonic.player.service.MediaFolderService;
+import org.airsonic.player.service.PersonalSettingsService;
+import org.airsonic.player.service.SecurityService;
+import org.airsonic.player.service.UserService;
+import org.airsonic.player.util.Util;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.MissingServletRequestParameterException;
+import org.springframework.web.bind.ServletRequestUtils;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.subsonic.restapi.Response;
+import org.subsonic.restapi.Users;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.util.List;
+import java.util.Optional;
+
+import static org.airsonic.player.security.RESTRequestParameterProcessingFilter.decrypt;
+import static org.springframework.web.bind.ServletRequestUtils.getBooleanParameter;
+import static org.springframework.web.bind.ServletRequestUtils.getIntParameter;
+import static org.springframework.web.bind.ServletRequestUtils.getRequiredStringParameter;
+import static org.springframework.web.bind.ServletRequestUtils.getStringParameter;
+
+@Controller
+@RequestMapping(value = {"/rest", "/ext"}, method = {RequestMethod.GET, RequestMethod.POST})
+public class SubsonicUserController {
+
+    @Autowired
+    private JAXBWriter jaxbWriter;
+    @Autowired
+    private UserService userService;
+    @Autowired
+    private SecurityService securityService;
+    @Autowired
+    private PersonalSettingsService personalSettingsService;
+    @Autowired
+    private MediaFolderService mediaFolderService;
+    @Autowired
+    private UserSettingsController userSettingsController;
+
+    @ExceptionHandler(MissingServletRequestParameterException.class)
+    public void handleMissingRequestParam(HttpServletRequest request,
+                                          HttpServletResponse response,
+                                          MissingServletRequestParameterException exception) {
+        error(request, response, SubsonicRESTController.ErrorCode.MISSING_PARAMETER,
+                "Required param (" + exception.getParameterName() + ") is missing");
+    }
+
+    @RequestMapping({"/changePassword", "/changePassword.view"})
+    public void changePassword(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        String username = getRequiredStringParameter(request, "username");
+        String password = decrypt(getRequiredStringParameter(request, "password"));
+
+        org.airsonic.player.domain.User authUser = securityService.getCurrentUser(request);
+
+        boolean allowed = authUser.isAdminRole()
+                || username.equals(authUser.getUsername()) && authUser.isSettingsRole();
+
+        if (!allowed) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, authUser.getUsername() + " is not authorized to change password for " + username);
+            return;
+        }
+
+        securityService.createAirsonicCredential(username, password, "Created via Subsonic REST API");
+
+        writeEmptyResponse(request, response);
+    }
+
+    @RequestMapping({"/getUser", "/getUser.view"})
+    public void getUser(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        String username = getRequiredStringParameter(request, "username");
+
+        org.airsonic.player.domain.User currentUser = securityService.getCurrentUser(request);
+        if (!username.equals(currentUser.getUsername()) && !currentUser.isAdminRole()) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, currentUser.getUsername() + " is not authorized to get details for other users.");
+            return;
+        }
+
+        org.airsonic.player.domain.User requestedUser = securityService.getUserByName(username);
+        if (requestedUser == null) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_FOUND, "No such user: " + username);
+            return;
+        }
+
+        Response res = createResponse();
+        res.setUser(createJaxbUser(requestedUser));
+        jaxbWriter.writeResponse(request, response, res);
+    }
+
+    @RequestMapping({"/getUsers", "/getUsers.view"})
+    public void getUsers(HttpServletRequest request, HttpServletResponse response) {
+        org.airsonic.player.domain.User currentUser = securityService.getCurrentUser(request);
+        if (!currentUser.isAdminRole()) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, currentUser.getUsername() + " is not authorized to get details for other users.");
+            return;
+        }
+
+        Users result = new Users();
+        for (org.airsonic.player.domain.User user : userService.getAllUsers()) {
+            result.getUser().add(createJaxbUser(user));
+        }
+
+        Response res = createResponse();
+        res.setUsers(result);
+        jaxbWriter.writeResponse(request, response, res);
+    }
+
+    private org.subsonic.restapi.User createJaxbUser(org.airsonic.player.domain.User user) {
+        UserSettings userSettings = personalSettingsService.getUserSettings(user.getUsername());
+
+        org.subsonic.restapi.User result = new org.subsonic.restapi.User();
+        result.setUsername(user.getUsername());
+        result.setEmail(user.getEmail());
+        result.setScrobblingEnabled(userSettings.getLastFmEnabled());
+        result.setAdminRole(user.isAdminRole());
+        result.setSettingsRole(user.isSettingsRole());
+        result.setDownloadRole(user.isDownloadRole());
+        result.setUploadRole(user.isUploadRole());
+        result.setPlaylistRole(true);  // Since 1.8.0
+        result.setCoverArtRole(user.isCoverArtRole());
+        result.setCommentRole(user.isCommentRole());
+        result.setPodcastRole(user.isPodcastRole());
+        result.setStreamRole(user.isStreamRole());
+        result.setJukeboxRole(user.isJukeboxRole());
+        result.setShareRole(user.isShareRole());
+        // currently this role isn't supported by airsonic
+        result.setVideoConversionRole(false);
+        // Useless
+        result.setAvatarLastChanged(null);
+
+        TranscodeScheme transcodeScheme = userSettings.getTranscodeScheme();
+        if (transcodeScheme != null && transcodeScheme != TranscodeScheme.OFF) {
+            result.setMaxBitRate(transcodeScheme.getMaxBitRate());
+        }
+
+        List<org.airsonic.player.domain.MusicFolder> musicFolders = mediaFolderService.getMusicFoldersForUser(user.getUsername());
+        for (org.airsonic.player.domain.MusicFolder musicFolder : musicFolders) {
+            result.getFolder().add(musicFolder.getId());
+        }
+        return result;
+    }
+
+    @RequestMapping({"/createUser", "/createUser.view"})
+    public void createUser(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        org.airsonic.player.domain.User user = securityService.getCurrentUser(request);
+        if (!user.isAdminRole()) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, user.getUsername() + " is not authorized to create new users.");
+            return;
+        }
+
+        UserSettingsCommand command = new UserSettingsCommand();
+        command.setUsername(getRequiredStringParameter(request, "username"));
+        command.setPassword(decrypt(getRequiredStringParameter(request, "password")));
+        command.setEmail(getRequiredStringParameter(request, "email"));
+        command.setLdapAuthenticated(getBooleanParameter(request, "ldapAuthenticated", false));
+        command.setAdminRole(getBooleanParameter(request, "adminRole", false));
+        command.setCommentRole(getBooleanParameter(request, "commentRole", false));
+        command.setCoverArtRole(getBooleanParameter(request, "coverArtRole", false));
+        command.setDownloadRole(getBooleanParameter(request, "downloadRole", false));
+        command.setStreamRole(getBooleanParameter(request, "streamRole", true));
+        command.setUploadRole(getBooleanParameter(request, "uploadRole", false));
+        command.setJukeboxRole(getBooleanParameter(request, "jukeboxRole", false));
+        command.setPodcastRole(getBooleanParameter(request, "podcastRole", false));
+        command.setSettingsRole(getBooleanParameter(request, "settingsRole", true));
+        command.setShareRole(getBooleanParameter(request, "shareRole", false));
+        command.setTranscodeSchemeName(TranscodeScheme.OFF.name());
+
+        int[] folderIds = ServletRequestUtils.getIntParameters(request, "musicFolderId");
+        if (folderIds.length == 0) {
+            folderIds = Util.toIntArray(org.airsonic.player.domain.MusicFolder.toIdList(mediaFolderService.getAllMusicFolders()));
+        }
+        command.setAllowedMusicFolderIds(folderIds);
+
+        userSettingsController.createUser(command);
+        writeEmptyResponse(request, response);
+    }
+
+    @RequestMapping({"/updateUser", "/updateUser.view"})
+    public void updateUser(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        org.airsonic.player.domain.User user = securityService.getCurrentUser(request);
+        if (!user.isAdminRole()) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, user.getUsername() + " is not authorized to update users.");
+            return;
+        }
+
+        String username = getRequiredStringParameter(request, "username");
+        org.airsonic.player.domain.User u = securityService.getUserByName(username);
+        UserSettings s = personalSettingsService.getUserSettings(username);
+
+        if (u == null) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_FOUND, "No such user: " + username);
+            return;
+        } else if (org.airsonic.player.domain.User.USERNAME_ADMIN.equals(username)) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, "Not allowed to change admin user");
+            return;
+        }
+
+        UserSettingsCommand command = new UserSettingsCommand();
+        command.setUsername(username);
+        command.setEmail(getStringParameter(request, "email", u.getEmail()));
+        command.setLdapAuthenticated(getBooleanParameter(request, "ldapAuthenticated", u.isLdapAuthenticated()));
+        command.setAdminRole(getBooleanParameter(request, "adminRole", u.isAdminRole()));
+        command.setCommentRole(getBooleanParameter(request, "commentRole", u.isCommentRole()));
+        command.setCoverArtRole(getBooleanParameter(request, "coverArtRole", u.isCoverArtRole()));
+        command.setDownloadRole(getBooleanParameter(request, "downloadRole", u.isDownloadRole()));
+        command.setStreamRole(getBooleanParameter(request, "streamRole", u.isStreamRole()));
+        command.setUploadRole(getBooleanParameter(request, "uploadRole", u.isUploadRole()));
+        command.setJukeboxRole(getBooleanParameter(request, "jukeboxRole", u.isJukeboxRole()));
+        command.setPodcastRole(getBooleanParameter(request, "podcastRole", u.isPodcastRole()));
+        command.setSettingsRole(getBooleanParameter(request, "settingsRole", u.isSettingsRole()));
+        command.setShareRole(getBooleanParameter(request, "shareRole", u.isShareRole()));
+
+        int maxBitRate = getIntParameter(request, "maxBitRate", s.getTranscodeScheme().getMaxBitRate());
+        command.setTranscodeSchemeName(Optional.ofNullable(TranscodeScheme.fromMaxBitRate(maxBitRate)).map(TranscodeScheme::name).orElse(null));
+
+        if (hasParameter(request, "password")) {
+            command.setPassword(decrypt(getRequiredStringParameter(request, "password")));
+            command.setPasswordChange(true);
+        }
+
+        int[] folderIds = ServletRequestUtils.getIntParameters(request, "musicFolderId");
+        if (folderIds.length == 0) {
+            folderIds = Util.toIntArray(org.airsonic.player.domain.MusicFolder.toIdList(mediaFolderService.getMusicFoldersForUser(username)));
+        }
+        command.setAllowedMusicFolderIds(folderIds);
+
+        userSettingsController.updateUser(command);
+        writeEmptyResponse(request, response);
+    }
+
+    private boolean hasParameter(HttpServletRequest request, String name) {
+        return request.getParameter(name) != null;
+    }
+
+    @RequestMapping({"/deleteUser", "/deleteUser.view"})
+    public void deleteUser(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        org.airsonic.player.domain.User user = securityService.getCurrentUser(request);
+        if (!user.isAdminRole()) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, user.getUsername() + " is not authorized to delete users.");
+            return;
+        }
+
+        String username = getRequiredStringParameter(request, "username");
+        try {
+            String currentUsername = securityService.getCurrentUsername(request);
+            securityService.deleteUser(username, currentUsername);
+        } catch (SecurityService.SelfDeletionException sde) {
+            error(request, response, SubsonicRESTController.ErrorCode.NOT_AUTHORIZED, sde.getMessage());
+            return;
+        }
+
+        writeEmptyResponse(request, response);
+    }
+
+    private Response createResponse() {
+        return jaxbWriter.createResponse(true);
+    }
+
+    private void writeEmptyResponse(HttpServletRequest request, HttpServletResponse response) {
+        jaxbWriter.writeResponse(request, response, createResponse());
+    }
+
+    private void error(HttpServletRequest request, HttpServletResponse response,
+                       SubsonicRESTController.ErrorCode code, String message) {
+        jaxbWriter.writeErrorResponse(request, response, code, message);
+    }
+}
